Consolidate dash import logic
